backend/requirements.txt
black>=22.10.0
celery>=5.2.7,<6.0.0
Django==4.0.2
django-cors-headers>=3.13.0,<4.0.0  # Only if actually using DRF
django-environ>=0.9.0
django-filter>=22.1,<23.0
djangorestframework>=3.14.0,<4.0.0  # Only if actually using DRF
firebase-admin>=6.0.1,<7.0.0  # Only if actually using Firebase
flake8>=6.0.0
isort>=5.10.1
Markdown<4.0.0
psycopg2-binary==2.9.5  # Only if used with PostgreSQL
redis==4.1.3  # Optional, only if using Celery, Channels, Redis cache or something Redis related
requests>=2.28.1
